pyproject.toml
[project]
name = "knowledge-service"
version = "0.1.0"
description = "Personal knowledge graph with Bayesian epistemics"
authors = [{name = "Arshad Ansari", email = "arshadansari27@gmail.com"}]
license = {text = "MIT"}
readme = "README.md"
keywords = ["knowledge-graph", "rdf", "bayesian", "sparql", "personal-knowledge"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pyoxigraph>=0.4.0",
    "problog>=2.2.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-httpx>=0.35.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://hikmahtechnologies.com"
Repository = "https://github.com/arshadansari27/knowledge-service"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[dependency-groups]
dev = [
    "pytest-httpx>=0.36.0",
]
